Changes to make it easier for the data release paper: - add solar angle calculation and field to the catalogue JSON - add the array configuration to the catalogue JSON - add the alphas to the flux density fit fields in the catalogue JSON - add the 64 MHz averaged real flux density closest to 5.5 GHz to the flux   density fit fields in the catalogue JSON
Made a script that went back to the catalogue JSON and filled these fields
in for the old data. The new generation script now adds these fields, along
with creating the averaged data files automatically.

The remove epoch script deals with the new information properly.

There is a new rename script that I made to deal with several epochs of
J0115-30B data being named as J0115-304.